backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.4
celery==5.3.6
celery[redis]==5.3.6
flower==2.0.1

# Scraping
playwright==1.44.0
playwright-stealth==1.0.6

# HTTP client
httpx==0.27.0

# Google Trends
pytrends==4.9.2

# AI / ML
scikit-learn==1.4.2
sentence-transformers==3.0.1
numpy==1.26.4
Pillow==10.3.0

# Telegram
python-telegram-bot==21.3

# Time series
prophet==1.1.5

# Caching
fastapi-cache2==0.2.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
